# book-network-api/src/main/java/com/luxus/book_network_api/book/BorrowedBookResponse.java
package com.luxus.book_network_api.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}